Add basic calculator functionality
The calculator is configured using an expression object to keep
track of operands and operators. The numbers are stored as strings
for display and concatenation and converted to numbers for
operations. Zero cannot be entered after a divide. Expression does
not evaluation without a secondary field in expression. Code needs
further cleanup and simplification.